+++ nvbench_helper/nvbench_helper/nvbench_helper.cuh
@@ -598,7 +598,7 @@ struct caching_allocator_t
 #if THRUST_DEVICE_SYSTEM == THRUST_DEVICE_SYSTEM_CUDA
   void* allocate_sync(size_t num_bytes, size_t)
   {
-    return allocate(num_bytes);
+    return allocate(static_cast<std::ptrdiff_t>(num_bytes));
   }
 
   void deallocate_sync(void* ptr, size_t num_bytes, size_t)
@@ -618,7 +618,7 @@ struct caching_allocator_t
     }
 
     value_type* result{};
-    auto free_block = free_blocks.find(num_bytes);
+    auto free_block = free_blocks.find(static_cast<std::ptrdiff_t>(num_bytes));
 
     if (free_block != free_blocks.end())
     {
